+++ apps/client/src/app/App.tsx
@@ -1,9 +1,9 @@
+import HomePage from '@pages/home/HomePage';
+import LoginPage from '@pages/login/LoginPage';
+import MapPage from '@pages/map/MapPage';
+import NotFound from '@pages/not-found/NotFound';
+import SignUpPage from '@pages/sign-up/SignUpPage';
 import { Navigate, Route, Routes } from 'react-router-dom';
-import HomePage from './pages/home/HomePage';
-import LoginPage from './pages/login/LoginPage';
-import MapPage from './pages/map/MapPage';
-import SignUpPage from './pages/signup/SignUpPage';
-import NotFound from './pages/not-found/NotFound';
 
 function App() {
   return (
@@ -13,7 +13,7 @@ function App() {
       <Route path='/map' element={<MapPage />} />
       <Route path='/home' element={<HomePage />} />
       <Route path='/login' element={<LoginPage />} />
-      <Route path='/signup' element={<SignUpPage />} />
+      <Route path='/sign-up' element={<SignUpPage />} />
     </Routes>
   );
 }
