+++ apps/client/src/app/pages/sign-up/components/ButtonSignUp.tsx
