+++ apps/client/src/app/pages/sign-up/components/LabelTextFile.tsx
