+++ apps/client/src/app/pages/sign-up/service/HandleSignUp.tsx
